backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
anthropic
pydantic
google-api-python-client
google-auth
